+++ tests/integration/cases/call_multi_args_comments/CSharp_call.cs
@@ -1,0 +1,9 @@
+using System;
+class Check {
+static object process(object ts = null, object start = null, object end = null) => null;
+    public static void Main() {
+process(1, 0, 3600);  // Jan 1 1970 00:00:00 - 01:00:00
+process(5, 0, 3600);  // Jan 1 1970 00:00:05 - 01:00:05
+process(2, 0, 5400);  // Jan 1 1970 00:00:02 - 01:30:02
+    }
+}