+++ tests/integration/cases/call_multi_args_comments/Java_call.java
@@ -1,0 +1,8 @@
+class Main {
+static Object process(Object... args) { return null; }
+    public static void main() {
+process(1, 0, 3600);  // Jan 1 1970 00:00:00 - 01:00:00
+process(5, 0, 3600);  // Jan 1 1970 00:00:05 - 01:00:05
+process(2, 0, 5400);  // Jan 1 1970 00:00:02 - 01:30:02
+    }
+}